+++ src/api/AxiosInstance.tsx
@@ -1,10 +1,10 @@
-import axios from "axios";
+import axios from 'axios';
 
-const BASE_URL = "https://fe-project-epigram-api.vercel.app/6-10";
+const BASE_URL = 'https://fe-project-epigram-api.vercel.app/6-10';
 
 export const instance = axios.create({
   baseURL: BASE_URL,
   headers: {
-    "Content-Type": "application/json",
+    'Content-Type': 'application/json',
   },
 });
