+++ src/types/modal.ts
@@ -1,0 +1,12 @@
+interface ButtonProps {
+  text: string;
+  onClick: () => void;
+  type?: 'primary' | 'secondary';
+}
+
+export interface ModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  message: string;
+  buttons: ButtonProps[];
+}
